整理钱包相关接口

+++ soho-extend/soho-wallet/soho-wallet-biz/src/main/java/work/soho/wallet/biz/service/WalletUserService.java
@@ -5,5 +5,5 @@ import com.baomidou.mybatisplus.extension.service.IService;
 
 public interface WalletUserService extends IService<WalletUser> {
     // 验证密码
-    Boolean verificationPayPassword(Integer userId, String payPassword);
+    Boolean verificationPayPassword(Long userId, String payPassword);
 }
